% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object_Utilities.R
\name{Add_Cell_Bender_Diff}
\alias{Add_Cell_Bender_Diff}
\title{Calculate and add differences post-cell bender analysis}
\usage{
Add_Cell_Bender_Diff(seurat_object, raw_assay_name, cell_bender_assay_name)
}
\arguments{
\item{seurat_object}{object name.}

\item{raw_assay_name}{name of the assay containing the raw data.}

\item{cell_bender_assay_name}{name of the assay containing the Cell Bender'ed data.}
}
\value{
Seurat object with 2 new columns in the meta.data slot.
}
\description{
Calculate the difference in features and UMIs per cell when both cell bender and raw assays are present.
}
\examples{
\dontrun{
object <- Add_Cell_Bender_Diff(seurat_object = obj, raw_assay_name = "RAW", cell_bender_assay_name = "RNA")
}

}
\concept{object_util}
