% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIGER_Utilities.R
\name{Add_Cell_Complexity_LIGER}
\alias{Add_Cell_Complexity_LIGER}
\title{Add Cell Complexity Value}
\usage{
Add_Cell_Complexity_LIGER(
  liger_object,
  meta_col_name = "log10GenesPerUMI",
  overwrite = FALSE
)
}
\arguments{
\item{liger_object}{object name.}

\item{meta_col_name}{name to use for new meta data column.  Default is "log10GenesPerUMI".}

\item{overwrite}{Logical.  Whether to overwrite existing an meta.data column.  Default is FALSE meaning that
function will abort if column with name provided to \code{meta_col_name} is present in meta.data slot.}
}
\value{
A LIGER Object
}
\description{
Add measure of cell complexity/novelty (log10PerUMI) for data QC.
}
\examples{
\dontrun{
object <- Add_Cell_Complexity_LIGER(liger_object = object)
}

}
\concept{liger_object_util}
