% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{as.anndata}
\alias{as.anndata}
\title{Convert objects to anndata objects}
\usage{
as.anndata(x, ...)
}
\arguments{
\item{x}{Seurat or LIGER object}

\item{...}{Arguments passed to other methods}
}
\description{
Convert objects to anndata objects
}
