% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{as.LIGER}
\alias{as.LIGER}
\title{Convert objects to LIGER objects}
\usage{
as.LIGER(x, ...)
}
\arguments{
\item{x}{An object to convert to class \code{LIGER}}

\item{...}{Arguments passed to other methods}
}
\description{
Convert objects to LIGER objects
}
