% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIGER_Utilities.R
\name{as.LIGER.Seurat}
\alias{as.LIGER.Seurat}
\title{Create liger object from one Seurat Object}
\usage{
\method{as.LIGER}{Seurat}(
  x,
  group.by = "orig.ident",
  assay = "RNA",
  remove_missing = FALSE,
  renormalize = TRUE,
  use_seurat_var_genes = FALSE,
  use_seurat_dimreduc = FALSE,
  reduction = NULL,
  keep_meta = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{group.by}{Variable in meta data which contains variable to split data by, (default is "orig.ident").}

\item{assay}{Assay containing raw data to use, (default is "RNA").}

\item{remove_missing}{logical, whether to remove missing genes with no counts when converting to
LIGER object (default is FALSE).}

\item{renormalize}{logical, whether to perform normalization after LIGER object creation (default is TRUE).}

\item{use_seurat_var_genes}{logical, whether to transfer variable features from Seurat object to
new LIGER object (default is FALSE).}

\item{use_seurat_dimreduc}{logical, whether to transfer dimensionality reduction coordinates from
Seurat to new LIGER object (default is FALSE).}

\item{reduction}{Name of Seurat reduction to transfer if \code{use_seurat_dimreduc = TRUE}.}

\item{keep_meta}{logical, whether to transfer columns in Seurat meta.data slot to LIGER cell.data
slot (default is TRUE).}

\item{verbose}{logical, whether to print status messages during object conversion (default is TRUE).}
}
\value{
liger object.
}
\description{
This function is part of generic \code{as.LIGER} for conversion of Seurat Objects.  Updated for Seurat5 compatibility and enhanced functionality.
}
\examples{
\dontrun{
liger_object <- as.LIGER(x = seurat_object)
}

}
\references{
modified and enhanced version of \code{rliger::seuratToLiger}.
}
\concept{object_conversion}
