% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics_Plotting.R
\name{CellBender_Diff_Plot}
\alias{CellBender_Diff_Plot}
\title{Plot Number of Cells/Nuclei per Sample}
\usage{
CellBender_Diff_Plot(
  feature_diff_df,
  label = TRUE,
  num_labels = 20,
  repel = TRUE,
  custom_labels = NULL,
  plot_title = "Raw Counts vs. Cell Bender Counts",
  x_axis_label = "Raw Data Counts (log10)",
  y_axis_label = "Cell Bender Counts (log10)",
  xnudge = 0,
  ynudge = 0,
  max.overlaps = 100,
  color = "red",
  fontface = "bold",
  ...
)
}
\arguments{
\item{feature_diff_df}{name of data.frame created using \code{\link[scCustomize]{CellBender_Feature_Diff}}}

\item{label}{logical, whether or not to label the features that have largest percent difference
between raw and CellBender counts (Default is TRUE).}

\item{num_labels}{Number of features to label if \code{label = TRUE}, (default is 20).}

\item{repel}{logical, whether to use geom_text_repel to create a nicely-repelled labels; this is
slow when a lot of points are being plotted. If using repel, set xnudge and ynudge to 0, (Default is TRUE).}

\item{custom_labels}{A custom set of features to label instead of the features most different between
raw and CellBender counts.}

\item{plot_title}{Plot title.}

\item{x_axis_label}{Label for x axis.}

\item{y_axis_label}{Label for y axis.}

\item{xnudge}{Amount to nudge X and Y coordinates of labels by}

\item{ynudge}{Amount to nudge X and Y coordinates of labels by}

\item{max.overlaps}{passed to \code{\link[ggrepel]{geom_text_repel}}, exclude text labels that
overlap too many things. Defaults to 100.}

\item{color}{Color to use for text labels.}

\item{fontface}{font face to use for text labels (“plain”, “bold”, “italic”, “bold.italic”).}

\item{...}{Extra parameters passed to \code{\link[ggrepel]{geom_text_repel}} through
\code{\link[Seurat]{LabelPoints}}.}
}
\value{
A ggplot object
}
\description{
Plot of total cell or nuclei number per sample grouped by another meta data variable.
}
\examples{
\dontrun{
# get cell bender differences data.frame
cb_stats <- CellBender_Feature_Diff(seurat_object - obj, raw_assay = "RAW",
cell_bender_assay = "RNA")

# plot
CellBender_Diff_Plot(feature_diff_df = cb_stats)
}

}
\concept{stats_plotting}
