% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{Cells_per_Sample}
\alias{Cells_per_Sample}
\title{Cells per Sample}
\usage{
Cells_per_Sample(seurat_object, sample_col = NULL)
}
\arguments{
\item{seurat_object}{Seurat object}

\item{sample_col}{column name in meta.data that contains sample ID information.  Default is NULL and
will use "orig.ident column}
}
\value{
A data.frame
}
\description{
Get data.frame containing the number of cells per sample.
}
\examples{
library(Seurat)
num_cells <- Cells_per_Sample(seurat_object = pbmc_small, sample_col = "orig.ident")

}
\concept{stats}
