% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Change_Delim_Suffix}
\alias{Change_Delim_Suffix}
\title{Change barcode suffix delimiter}
\usage{
Change_Delim_Suffix(data, current_delim, new_delim)
}
\arguments{
\item{data}{Either matrix/data.frame or list of matrices/data.frames with the cell barcodes in the column names.}

\item{current_delim}{a single value of current delimiter.}

\item{new_delim}{a single value of new delimiter desired.}
}
\description{
Change barcode suffix delimiter from list of data.frames/matrices or single data.frame/matrix
}
\examples{
\dontrun{
dge_matrix <- Change_Delim_Suffix(data = dge_matrix, current_delim = ".", new_delim = "-")
}

}
\concept{helper_util}
