% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Color_Palettes.R
\name{ColorBlind_Pal}
\alias{ColorBlind_Pal}
\title{Color Universal Design Short Palette}
\usage{
ColorBlind_Pal()
}
\value{
modified/reordered color palette (8 colors) based on ditto-seq
}
\description{
Shortcut ta a modified 8 color palette based on Color Universal Design (CUD) colorblindness friendly palette.
}
\examples{
cols <- ColorBlind_Pal()
PalettePlot(pal = cols)

}
\references{
palette is slightly modified version of the Color Universal Design (CUD) colorblindness
friendly palette \url{https://jfly.uni-koeln.de/color/}.
}
\concept{palettes}
