% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Copy_To_GCP}
\alias{Copy_To_GCP}
\title{Copy folder to GCP bucket from R Console}
\usage{
Copy_To_GCP(folder_file_path, gcp_bucket_path)
}
\arguments{
\item{folder_file_path}{folder to be copied to GCP bucket.}

\item{gcp_bucket_path}{GCP bucket path to copy to files.}
}
\description{
Run command from R console without moving to terminal to copy folder to GCP bucket
}
\examples{
\dontrun{
Copy_To_GCP(folder_file_path = "plots/", gcp_bucket_path = "gs://bucket_name_and_folder_path")
}

}
\concept{organization_util}
