% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Create_Cluster_Annotation_File}
\alias{Create_Cluster_Annotation_File}
\title{Create cluster annotation csv file}
\usage{
Create_Cluster_Annotation_File(
  file_path = NULL,
  file_name = "cluster_annotation"
)
}
\arguments{
\item{file_path}{path to directory to save file.  Default is current working directory.}

\item{file_name}{name to use for annotation file.  Function automatically adds file type ".csv" suffix.
Default is "cluster_annotation".}
}
\value{
No value returned.  Creates .csv file.
}
\description{
create annotation file
}
\examples{
\dontrun{
Create_Cluster_Annotation_File(file_path = "cluster_annotation_folder_name")
}

}
\concept{marker_annotation_util}
