% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Color_Palettes.R
\name{Dark2_Pal}
\alias{Dark2_Pal}
\title{Dark2 Palette}
\usage{
Dark2_Pal()
}
\value{
"Dark2" color palette (8 colors)
}
\description{
Shortcut to Dark2 color palette from RColorBrewer (8 Colors)
}
\examples{
\dontrun{
cols <- Dark2_Pal()
}

}
\references{
Dark2 palette from RColorBrewer being called through paletteer.
See RColorBrewer for more info on palettes
\url{https://CRAN.R-project.org/package=RColorBrewer}
}
\concept{palettes}
