% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIGER_Utilities.R
\name{Dataset_Size_LIGER}
\alias{Dataset_Size_LIGER}
\title{Check size of LIGER datasets}
\usage{
Dataset_Size_LIGER(liger_object, other_meta = NULL)
}
\arguments{
\item{liger_object}{LIGER object name.}

\item{other_meta}{other meta data to include in returned data.frame}
}
\value{
data.frame containing cells
}
\description{
Returns size (number of cells) in each dataset within liger object along with other desired meta data.
}
\examples{
\dontrun{
cells_per_dataset <- Dataset_Size_LIGER(liger_object = object, other_meta = c("Age", "Sex"))
}

}
\concept{liger_object_util}
