% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{Add_Top_Gene_Pct_Seurat}
\alias{Add_Top_Gene_Pct_Seurat}
\title{Deprecated functions \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
Add_Top_Gene_Pct_Seurat(...)
}
\description{
Use \code{\link[=Add_Top_Gene_Pct]{Add_Top_Gene_Pct()}} instead of \code{Add_Top_Gene_Pct_Seurat()}.
}
\keyword{internal}
