% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Color_Palettes.R
\name{DiscretePalette_scCustomize}
\alias{DiscretePalette_scCustomize}
\title{Discrete colour palettes from the pals package}
\usage{
DiscretePalette_scCustomize(
  num_colors,
  palette = NULL,
  shuffle_pal = FALSE,
  seed = 123
)
}
\arguments{
\item{num_colors}{Number of colors to be generated.}

\item{palette}{Options are
"alphabet", "alphabet2", "glasbey", "polychrome", "stepped", "ditto_seq", "varibow".
Can be omitted and the function will use the one based on the requested n.}

\item{shuffle_pal}{randomly shuffle the outputted palette.  Most useful for \code{varibow} palette as
that is normally an ordered palette.}

\item{seed}{random seed for the palette shuffle.  Default = 123.}
}
\value{
A vector of colors
}
\description{
These are included here because pals depends on a number of compiled
packages, and this can lead to increases in run time for Travis,
and generally should be avoided when possible.
}
\references{
This function uses the paletteer package (https://github.com/EmilHvitfeldt/paletteer) to
provide simplified access to color palettes from many different R package sources while
minimizing scCustomize current and future dependencies.

The following packages & palettes are called by paletteer (see individual packages for
palette references/citations):
\enumerate{
\item pals \url{https://cran.r-project.org/web/packages/pals/index.html}
\itemize{
\item alphabet, alphabet2, glasbey, polychrome, and stepped.
}
}

Function name and implementation modified from Seurat (Licence: GPL-3).
\url{https://github.com/satijalab/seurat}
}
\concept{palettes}
