% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object_Utilities.R
\name{Downsample_Seurat}
\alias{Downsample_Seurat}
\title{Downsample Seurat Object}
\usage{
Downsample_Seurat(
  seurat_object,
  num_cells,
  group.by = NULL,
  allow_lower = FALSE,
  seed = 123
)
}
\arguments{
\item{seurat_object}{name of Seurat object}

\item{num_cells}{number of cells per ident to use in down-sampling.  This value must be less than or
equal to the size of ident with fewest cells.  Alternatively, can set to "min" which will
use the maximum number of barcodes based on size of smallest group.}

\item{group.by}{The ident to use to group cells.  Default is NULL which use current active.ident.  .}

\item{allow_lower}{logical, if number of cells in identity is lower than \code{num_cells} keep the
maximum number of cells, default is FALSE.  If FALSE will report error message if \code{num_cells} is
too high, if TRUE will subset cells with more than \code{num_cells} to that value and those with less
than \code{num_cells} will not be downsampled.}

\item{seed}{random seed to use for downsampling.  Default is 123.}
}
\value{
Seurat object with randomly downsampled cells.
}
\description{
Randomly downsample Seurat object by given number of cells per group (or by group with smallest number of cells)
}
\examples{
\dontrun{
# Downsample specific number of cells per group
obj_sub <- Downsample_Seurat(seurat_object = obj, num_cells = 1000)

# Downsample specific number of cells per group but allow groups to have fewer cells
# if they don't have number provided
obj_sub <- Downsample_Seurat(seurat_object = obj, num_cells = 1000, allow_lower = TRUE)

# Downsample by number of cells in the smallest group
obj_sub <- Downsample_Seurat(seurat_object = obj, num_cells = "min")
}

}
