% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{ensembl_exAM_list}
\alias{ensembl_exAM_list}
\title{Immediate Early Gene (IEG) gene lists}
\format{
A list of three vectors
\describe{
\item{Mus_musculus_exAM_union_ensembl}{Ensembl ID for exAM genes from source publication (see below)}
\item{Homo_sapiens_exAM_union_ensembl}{Human Ensembl ID for exAM genes for homologous genes from mouse gene list}
\item{Homo_sapiens_exAM_micro_ensembl}{Human Ensembl ID for exAM genes for human microglia list}

}
}
\source{
Gene list is from: SI Table 22 Marsh et al., 2022 (Nature Neuroscience) from \doi{10.1038/s41593-022-01022-8}.
See data-raw directory for scripts used to create gene list.
}
\usage{
ensembl_exAM_list
}
\description{
Ensembl IDs for immediate early genes (Ensembl version 112; 4/29/2024)
}
\concept{data}
\keyword{datasets}
