% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{ensembl_hemo_id}
\alias{ensembl_hemo_id}
\title{Ensembl Hemo IDs}
\format{
A list of six vectors
\describe{
\item{Mus_musculus_hemo_ensembl}{Ensembl IDs for mouse hemoglobin genes}
\item{Homo_sapiens_hemo_ensembl}{Ensembl IDs for human hemoglobin genes}
\item{Danio_rerio_hemo_ensembl}{Ensembl IDs for zebrafish hemoglobin genes}
\item{Rattus_norvegicus_hemo_ensembl}{Ensembl IDs for rat hemoglobin genes}
\item{Drosophila_melanogaster_hemo_ensembl}{Ensembl IDs for fly hemoglobin genes}
\item{Macaca_mulatta_hemo_ensembl}{Ensembl IDs for macaque hemoglobin genes}
\item{Gallus_gallus_ribo_ensembl}{Ensembl IDs for chicken hemoglobin genes}
}
}
\source{
See data-raw directory for scripts used to create gene list.
}
\usage{
ensembl_hemo_id
}
\description{
A list of ensembl ids for hemoglobin genes (Ensembl version 112; 4/29/2024)
}
\concept{data}
\keyword{datasets}
