% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{ensembl_ieg_list}
\alias{ensembl_ieg_list}
\title{Immediate Early Gene (IEG) gene lists}
\format{
A list of seven vectors
\describe{
\item{Mus_musculus_IEGs}{Ensembl IDs for IEGs from source publication (see below)}
\item{Homo_sapiens_IEGs}{Ensembl IDs for homologous genes from mouse gene list}

}
}
\source{
Mouse gene list is from: SI Table 4 from \doi{10.1016/j.neuron.2017.09.026}.  Human
gene list was compiled by first creating homologous gene list using biomaRt and then adding some manually curated
homologs according to HGNC.  See data-raw directory for scripts used to create gene list.
}
\usage{
ensembl_ieg_list
}
\description{
Ensembl IDs for immediate early genes
}
\concept{data}
\keyword{datasets}
