% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{ensembl_lncRNA_id}
\alias{ensembl_lncRNA_id}
\title{Ensembl lncRNA IDs}
\format{
A list of seven vectors
\describe{
\item{Mus_musculus_lncRNA_ensembl}{Ensembl IDs for mouse lncRNA genes}
\item{Homo_sapiens_lncRNA_ensembl}{Ensembl IDs for human lncRNA genes}
\item{Callithrix_jacchus_lncRNA_ensembl}{Ensembl IDs for marmoset lncRNA genes}
\item{Danio_rerio_lncRNA_ensembl}{Ensembl IDs for zebrafish lncRNA genes}
\item{Rattus_norvegicus_lncRNA_ensembl}{Ensembl IDs for rat lncRNA genes}
\item{Macaca_mulatta_lncRNA_ensembl}{Ensembl IDs for macaque lncRNA genes}
\item{Gallus_gallus_lncRNA_ensembl}{Ensembl IDs for chicken lncRNA genes}
}
}
\source{
See data-raw directory for scripts used to create gene list.
}
\usage{
ensembl_lncRNA_id
}
\description{
A list of ensembl ids for lncRNA genes (Ensembl version 113; 04/08/2025)
}
\concept{data}
\keyword{datasets}
