% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{ensembl_malat1_list}
\alias{ensembl_malat1_list}
\title{MALAT1 gene lists}
\format{
A list of seven vectors
\describe{
\item{Mus_musculus_MALAT1_ensembl}{Ensembl ID for mouse Malat1}
\item{Homo_sapiens_MALAT1_ensembl}{Ensembl ID for human MALAT1}

}
}
\source{
See data-raw directory for scripts used to create gene list.
}
\usage{
ensembl_malat1_list
}
\description{
Ensembl IDs for MALAT1 (Ensembl version 112; 4/29/2024)
}
\concept{data}
\keyword{datasets}
