% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{ensembl_mito_id}
\alias{ensembl_mito_id}
\title{Ensembl Mito IDs}
\format{
A list of six vectors
\describe{
\item{Mus_musculus_mito_ensembl}{Ensembl IDs for mouse mitochondrial genes}
\item{Homo_sapiens_mito_ensembl}{Ensembl IDs for human mitochondrial genes}
\item{Danio_rerio_mito_ensembl}{Ensembl IDs for zebrafish mitochondrial genes}
\item{Rattus_norvegicus_mito_ensembl}{Ensembl IDs for rat mitochondrial genes}
\item{Drosophila_melanogaster_mito_ensembl}{Ensembl IDs for fly mitochondrial genes}
\item{Macaca_mulatta_mito_ensembl}{Ensembl IDs for macaque mitochondrial genes}

}
}
\usage{
ensembl_mito_id
}
\description{
A list of ensembl ids for mitochondrial genes (Ensembl version 105)
}
\concept{data}
\keyword{datasets}
