% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{ensembl_ribo_id}
\alias{ensembl_ribo_id}
\title{Ensembl Ribo IDs}
\format{
A list of seven vectors
\describe{
\item{Mus_musculus_ribo_ensembl}{Ensembl IDs for mouse ribosomal genes}
\item{Homo_sapiens_ribo_ensembl}{Ensembl IDs for human ribosomal genes}
\item{Callithrix_jacchus_ribo_ensembl}{Ensembl IDs for marmoset ribosomal genes}
\item{Danio_rerio_ribo_ensembl}{Ensembl IDs for zebrafish ribosomal genes}
\item{Rattus_norvegicus_ribo_ensembl}{Ensembl IDs for rat ribosomal genes}
\item{Drosophila_melanogaster_ribo_ensembl}{Ensembl IDs for fly ribosomal genes}
\item{Macaca_mulatta_ribo_ensembl}{Ensembl IDs for macaque ribosomal genes}

}
}
\usage{
ensembl_ribo_id
}
\description{
A list of ensembl ids for ribosomal genes (Ensembl version 105)
}
\concept{data}
\keyword{datasets}
