% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{exAM_gene_list}
\alias{exAM_gene_list}
\title{exAM gene lists}
\format{
A list of three vectors
\describe{
\item{Mus_musculus_exAM_union}{Gene symbols for exAM genes from source publication (see below)}
\item{Homo_sapiens_exAM_union}{Human gene symbols for homologous genes from mouse gene list}
\item{Homo_sapiens_exAM_micro}{Human gene symbols for human microglia list}

}
}
\source{
Gene list is from: SI Table 22 Marsh et al., 2022 (Nature Neuroscience) from \doi{10.1038/s41593-022-01022-8}.
See data-raw directory for scripts used to create gene list.
}
\usage{
exAM_gene_list
}
\description{
Gene symbols for exAM genes
}
\concept{data}
\keyword{datasets}
