% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object_Utilities.R
\name{Extract_Sample_Meta}
\alias{Extract_Sample_Meta}
\title{Extract sample level meta.data}
\usage{
Extract_Sample_Meta(
  object,
  sample_name = "orig.ident",
  variables_include = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{sample_name}{meta.data column to use as sample.  Output data.frame will contain one row per
level or unique value in this variable.}

\item{variables_include}{\verb{@meta.data} columns to keep in final data.frame.  All other columns will
be discarded.  Default is NULL.}

\item{variable_exclude}{columns to discard in final data.frame.  Many cell level columns are
irrelevant at the sample level (e.g., nFeature_RNA, percent_mito).  Default is to
discard nFeature_ASSAY(s), nCount_ASSAY(s), percent_mito, percent_ribo, percent_mito_ribo.  \emph{If
sample level median values are desired for these type of variables the output of this function
can be joined with output of \code{\link[scCustomize]{Median_Stats}}.}}
}
\value{
Returns a dataframe with one row per \code{sample_name}.
}
\description{
Returns a by identity meta.data data.frame with one row per sample.  Useful for downstream
quick view of sample breakdown, meta data table creation, and/or use in pseudobulk analysis
}
\examples{
\dontrun{
sample_meta <- AggregateMeta(object = pbmc_small)

Select meta.data columns to keep in ouput
sample_meta <- AggregateMeta(object = ifnb, variables_include = c("orig.ident", "stim"))
}

}
