% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIGER_Plotting.R
\name{Factor_Cor_Plot}
\alias{Factor_Cor_Plot}
\title{Factor Correlation Plot}
\usage{
Factor_Cor_Plot(
  liger_object,
  colors_use = NULL,
  label = FALSE,
  label_threshold = 0.5,
  label_size = 5,
  plot_title = NULL,
  plot_type = "full",
  positive_only = FALSE,
  x_lab_rotate = TRUE,
  cluster = TRUE,
  cluster_rect = FALSE,
  cluster_rect_num = NULL,
  cluster_rect_col = "white"
)
}
\arguments{
\item{liger_object}{liger object.}

\item{colors_use}{Color palette to use for correlation values.  Default is \code{viridis}.
Users can also supply vector of 3 colors (low, mid, high).}

\item{label}{logical, whether to add correlation values to plot result.}

\item{label_threshold}{threshold for adding correlation values if \code{label = TRUE}.  Default
is 0.5.}

\item{label_size}{size of correlation labels}

\item{plot_title}{Plot title.}

\item{plot_type}{Controls plotting full matrix, or just the upper or lower triangles.
Accepted values are: "full" (default), "upper", or "lower".}

\item{x_lab_rotate}{logical, whether to rotate the axes labels on the x-axis.  Default is TRUE}
}
\value{
A ggplot object
}
\description{
Plot positive correlations between gene loadings across \code{W} factor matrix in liger object.
Any negative correlations are set to NA and NA values set to bottom color of color gradient.
}
\examples{
\dontrun{
Factor_Cor_Plot(liger_object = obj)
}

}
\concept{liger_plotting}
