% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Fetch_Meta}
\alias{Fetch_Meta}
\title{Get meta data from object}
\usage{
Fetch_Meta(object)
}
\arguments{
\item{object}{list of matrices to merge.}
}
\value{
A data.frame
}
\description{
Quick function to properly pull meta.data from objects.
}
\examples{
\dontrun{
data_list <- Read10X_GEO(...)
merged <- Merge_Sparse_Data_All(matrix_list = data_list, add_cell_ids = names(data_list),
prefix = TRUE, cell_id_delimiter = "_")
}

}
\concept{helper_util}
