% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIGER_Utilities.R
\name{Find_Factor_Cor}
\alias{Find_Factor_Cor}
\title{Find Factor Correlations}
\usage{
Find_Factor_Cor(liger_object)
}
\arguments{
\item{liger_object}{LIGER object name.}
}
\value{
correlation matrix
}
\description{
Calculate correlations between gene loadings for all factors in liger object.
}
\examples{
\dontrun{
factor_correlations <- Find_Factor_Cor(liger_object = object)
}

}
\concept{liger_object_util}
