% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_Utilities_Internal.R
\name{get_plot_limits}
\alias{get_plot_limits}
\title{Get plot limits from ggplot2 object}
\usage{
get_plot_limits(plot)
}
\arguments{
\item{plot}{ggplot2 object to get axis limits from}
}
\value{
list with x and y axis min and max values
}
\description{
Get plot limits from ggplot2 object
}
\references{
Function from StackOverflow \url{https://stackoverflow.com/a/40304848}
}
