% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Color_Palettes.R
\name{Hue_Pal}
\alias{Hue_Pal}
\title{Hue_Pal}
\usage{
Hue_Pal(num_colors)
}
\arguments{
\item{num_colors}{number of colors to return in palette.}
}
\value{
hue color palette (as many colors as desired)
}
\description{
Shortcut to hue_pal to return to ggplot2 defaults if user desires, from scales package.
}
\examples{
\dontrun{
cols <- Hue_Pal(num_colors = 8)
}

}
\concept{palettes}
