% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_Iterative_Plotting.R
\name{Iterate_VlnPlot}
\alias{Iterate_VlnPlot}
\title{Iterative Plotting of Gene Lists using VlnPlots}
\usage{
Iterate_VlnPlot(
  seurat_object,
  gene_list,
  colors_use = NULL,
  pt.size = NULL,
  file_path = NULL,
  file_name = NULL,
  file_type = NULL,
  single_pdf = FALSE,
  dpi = 600,
  ...
)
}
\arguments{
\item{seurat_object}{Seurat object name.}

\item{gene_list}{list of genes to plot.}

\item{colors_use}{color scheme to use.}

\item{pt.size}{point size for plotting.}

\item{file_name}{name suffix and file extension.}

\item{file_type}{File type to save output as.  Must be one of following: ".pdf", ".png", ".tiff", ".jpeg", or ".svg".}

\item{single_pdf}{saves all plots to single PDF file (default = FALSE).  `file_type`` must be .pdf.}

\item{dpi}{dpi for image saving.}

\item{...}{Extra parameters passed to \code{\link[Seurat]{VlnPlot}}.}

\item{file_path/prefix}{directory file path and/or file name prefix.  Defaults to current wd.}
}
\description{
Create and Save plots for Gene list with Single Command
}
\examples{
\dontrun{
Iterate_VlnPlot(seurat_object = object, gene_list = DEG_list, colors = color_list, file_path = "plots/",
file_name = "_vln", file_type = ".jpg", dpi = 600)
}

}
\concept{iterative_plotting}
