% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Color_Palettes.R
\name{JCO_Four}
\alias{JCO_Four}
\title{Four Color Palette (JCO)}
\usage{
JCO_Four()
}
\value{
4 color palette from the JCO ggsci palette
}
\description{
Shortcut to a specific JCO 4 color palette from ggsci package.
}
\examples{
\dontrun{
cols <- JCO_Four()
}

}
\references{
Selection of colors from the JCO palette from ggsci being called through paletteer.
See ggsci for more info on palettes
\url{https://CRAN.R-project.org/package=ggsci}
}
\concept{palettes}
