% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIGER_Utilities.R
\name{LIGER_DimReduc}
\alias{LIGER_DimReduc}
\title{Extract dimensionality reduction coordinates from Liger object}
\usage{
LIGER_DimReduc(liger_object, reduction = NULL)
}
\arguments{
\item{liger_object}{LIGER object name.}

\item{reduction}{name of dimensionality reduction stored in cellMeta slot.  Default is
"UMAP")}
}
\value{
dimensionality reduction coordinates in 2 column format
}
\description{
Extract data.frame containing dimensionality reduction coordinates from new format of
Liger objects
}
\examples{
\dontrun{
umap_coords <- LIGER_DimReduc(liger_object = object)
}

}
\concept{liger_object_util}
