% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{lncRNA_gene_list}
\alias{lncRNA_gene_list}
\title{lncRNA gene list}
\format{
A list of six vectors
\describe{
\item{Mus_musculus_lncRNA}{Ensembl IDs for mouse lncRNA genes}
\item{Homo_sapiens_lncRNA}{Ensembl IDs for human lncRNA genes}
\item{Danio_rerio_lncRNA}{Ensembl IDs for zebrafish lncRNA genes}
\item{Rattus_norvegicus_lncRNA}{Ensembl IDs for rat lncRNA genes}
\item{Macaca_mulatta_lncRNA}{Ensembl IDs for macaque lncRNA genes}
\item{Gallus_gallus_lncRNA}{Ensembl IDs for chicken lncRNA genes}
}
}
\source{
See data-raw directory for scripts used to create gene list.
}
\usage{
lncRNA_gene_list
}
\description{
A list of gene symbol ids for lncRNA genes (Ensembl version 113; 04/08/2025)
}
\concept{data}
\keyword{datasets}
