% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Meta_Numeric}
\alias{Meta_Numeric}
\title{Check if meta data columns are numeric}
\usage{
Meta_Numeric(data)
}
\arguments{
\item{data}{a data.frame contain meta.data.}
}
\value{
vector of meta data columns that are numeric.
}
\description{
Check if any present meta data columns are numeric and returns vector of valid numeric columns.
Issues warning message if any columns not in numeric form.
}
\examples{
\dontrun{
numeric_meta_columns <- Meta_Numeric(data = meta_data)
}

}
\concept{helper_util}
