% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Color_Palettes.R
\name{NavyAndOrange}
\alias{NavyAndOrange}
\title{Navy and Orange Dual Color Palette}
\usage{
NavyAndOrange(flip_order = FALSE)
}
\arguments{
\item{flip_order}{Whether to flip the order of colors.}
}
\value{
Navy orange palette
}
\description{
Shortcut to navy orange color plot
}
\examples{
cols <- NavyAndOrange()
PalettePlot(palette = cols)

}
\concept{palettes}
