% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_Utilities.R
\name{PC_Plotting}
\alias{PC_Plotting}
\title{PC Plots}
\usage{
PC_Plotting(seurat_object, dim_number)
}
\arguments{
\item{seurat_object}{Seurat Object.}

\item{dims_number}{A single dim to plot (integer).}
}
\value{
A plot of PC heatmap and gene loadings for single
}
\description{
Plot PC Heatmaps and Dim Loadings for exploratory analysis.  Plots a single Heatmap and Gene Loading Plot.
Used for PC_Loading_Plots function.
}
\examples{
\dontrun{
PC_Plotting(seurat_object = seurat, dims_plot = 25, "plots/")
}

}
\seealso{
\code{\link[Seurat]{PCHeatmap}} and \code{\link[Seurat]{VizDimLoadings}}
}
\concept{seurat_plotting}
