% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics_Plotting.R
\name{Plot_Pie_Proportions}
\alias{Plot_Pie_Proportions}
\title{Cell Proportion Pie Chart}
\usage{
Plot_Pie_Proportions(
  seurat_object,
  group_by_var = "Total",
  colors_use = NULL,
  ggplot_default_colors = FALSE,
  color_seed = 123
)
}
\arguments{
\item{seurat_object}{Seurat object name.}

\item{group_by_var}{meta data column to classify samples (default = "Total"; totals across all samples).}

\item{colors_use}{color palette to use for plotting.}

\item{ggplot_default_colors}{logical.  If \code{colors_use = NULL}, Whether or not to return plot using
default ggplot2 "hue" palette instead of default "polychrome" or "varibow" palettes.}

\item{color_seed}{random seed for the "varibow" palette shuffle if \code{colors_use = NULL} and number of
groups plotted is greater than 36.  Default = 123.}
}
\value{
ggplot2 or patchwork object
}
\description{
Plots the proportion of cells belonging to each identity in \code{active.ident} of Seurat object.
Can plot either the totals or split by a variable in \code{meta.data}.
}
\examples{
#' library(Seurat)
Plot_Pie_Proportions(seurat_object = pbmc_small)

}
