% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_&_Write_Data.R
\name{Pull_Directory_List}
\alias{Pull_Directory_List}
\title{Pull Directory List}
\usage{
Pull_Directory_List(base_path)
}
\arguments{
\item{base_path}{path to the parent directory which contains all of the subdirectories of interest.}
}
\value{
A vector of sub-directories within \code{base_path}.
}
\description{
Enables easy listing of all sub-directories for use as input library lists in Read10X multi functions.
}
\examples{
\dontrun{
data_dir <- 'path/to/data/directory'
library_list <- Pull_Directory_List(base_path = data_dir)
}

}
\concept{read_&_write}
