% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_&_Write_Data.R
\name{Read_CellBender_h5_Multi_File}
\alias{Read_CellBender_h5_Multi_File}
\title{Load CellBender h5 matrices (corrected) from multiple files}
\usage{
Read_CellBender_h5_Multi_File(
  data_dir = NULL,
  filtered_h5 = TRUE,
  custom_name = NULL,
  sample_list = NULL,
  sample_names = NULL,
  parallel = FALSE,
  num_cores = NULL,
  ...
)
}
\arguments{
\item{data_dir}{Directory containing the .h5 files output by CellBender.}

\item{filtered_h5}{BLANK}

\item{custom_name}{BLANK}

\item{sample_list}{a vector of sample names if only specific samples are desired.  If \code{NULL} will
read in all files within \code{data_dir} directory.}

\item{sample_names}{a set of sample names to use for each sample entry in returned list.  If \code{NULL} will
set names to the subdirectory name of each sample.}

\item{parallel}{logical (default FALSE) whether or not to use multi core processing to read in matrices.}

\item{num_cores}{how many cores to use for parallel processing.}

\item{...}{Extra parameters passed to \code{\link[scCustomize]{Read_CellBender_h5_Mat}}.}
}
\value{
list of sparse matrices
}
\description{
Extract sparse matrix with corrected counts from CellBender h5 output file across multiple samples
within the same directory.
}
\examples{
\dontrun{
base_path <- 'path/to/data/directory'
mat_list <- Read_CellBender_h5_Multi_File(data_dir = base_path)
}

}
\concept{read_&_write}
