% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_&_Write_Data.R
\name{Read_Metrics_CellBender}
\alias{Read_Metrics_CellBender}
\title{Read Overall Statistics from CellBender}
\usage{
Read_Metrics_CellBender(base_path, lib_list = NULL, lib_names = NULL)
}
\arguments{
\item{base_path}{path to the parent directory which contains all of the sub-directories of interest or
path to single metrics csv file.}

\item{lib_list}{a list of sample names (matching directory names) to import.  If \code{NULL} will read
in all samples in parent directory.}

\item{lib_names}{a set of sample names to use for each sample.  If \code{NULL} will set names to the
directory name of each sample.}
}
\value{
A data frame with sample metrics from CellBender.
}
\description{
Get data.frame with all metrics from the CellBender \code{remove-background} analysis.
}
\examples{
\dontrun{
CB_metrics <- Read_Metrics_CellBender(base_path = "/path/to/directories")
}

}
\concept{read_&_write}
