% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_Write_Data.R
\name{Read_Velocity}
\alias{Read_Velocity}
\title{Load RNA Velocity results}
\usage{
Read_Velocity(loom_file)
}
\arguments{
\item{loom_file}{}
}
\value{
list of sparse matrices containing spliced, unspliced, and ambiguous counts
}
\description{
Wrapper function to
}
\examples{
\dontrun{
velo_res <- Read_Velocity(loom_file = "PATH/sample01.loom")
}
}
\references{
Function is updated from \code{SeuratWrappers::ReadVelocity}
\url{https://github.com/satijalab/seurat-wrappers/blob/master/R/velocity.R} (License: GPL-3).
Function included in scCustomize to avoid installation issues related to other aspects of
velocyto.R package which \code{SeuratWrappers::ReadVelocity} requires.
}
