% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{Run_Module_Sig}
\alias{Run_Module_Sig}
\title{Calculate Enrichment Cutoff for Module Scores}
\usage{
Run_Module_Sig(
  seurat_object,
  score_name,
  num_features,
  downsample_cell_num = 2000,
  allow_lower = FALSE,
  plot_mad_cutoffs = TRUE,
  plot_enrichment = TRUE,
  colors_use_enrichment = NULL,
  ggplot_default_colors = FALSE,
  color_seed = 123,
  return_plots = FALSE,
  seed = 123,
  verbose = TRUE
)
}
\arguments{
\item{seurat_object}{seurat object.}

\item{score_name}{name of column in object meta.data containing the score to test against.}

\item{num_features}{number of features in list used to create \code{score_name}.}

\item{downsample_cell_num}{number of cells to downsample each identity to.  Must be less than
number of cells in smallest identity unless \code{allow_lower = TRUE}.}

\item{allow_lower}{logical, allow identities with less cells than \code{downsample_cell_num} to be remain
unsampled.}

\item{plot_mad_cutoffs}{logical, whether to print mad_cutoff plots, default is TRUE.}

\item{plot_enrichment}{logical, whether to print enrichment percentage plots, default is TRUE.}

\item{colors_use_enrichment}{colors to use for each identity plotted in enrichment percentage plot.}

\item{ggplot_default_colors}{logical.  If \code{colors_use = NULL}, Whether or not to return plot using
default ggplot2 "hue" palette instead of default "polychrome" or "varibow" palettes.}

\item{color_seed}{random seed for the "varibow" palette shuffle if \code{colors_use = NULL} and number of
groups plotted is greater than 36.  Default = 123.}

\item{return_plots}{logical, whether to save plots to \verb{@misc} slot of object.}

\item{seed}{random seed to use for both object downsampling and module scoring.}

\item{verbose}{logical, whether or not to print messages while running function.}
}
\value{
seurat object with cutoff threshold value and percent enriched data.frame in \verb{@misc} slot.
If \code{return_plots = TRUE} also returns plots in the \verb{@misc} slot.
}
\description{
Calculate enrichment cutoff for module score to determine whether score is "enriched" in a given cell.
Also returns plots for cutoff determination and cell enrichment.
See details below for more information on usage and how the enrichment cutoff is calculated.
}
\details{
This function is based on method used in Marsh et al., (2022).  In short it uses random gene list
scoring to determine an outlier threshold.  Any score above that threshold for desired score is
then considered enriched,

User supplies \code{num_features} that indicates the number of features used in creation of the score
to test against.
User also supplies \code{downsample_cell_num} to downsample the original object.  Speeds up function and
ensures that all identities are equally represented.
Function will then:
\enumerate{
\item Create 1000 random gene lists of size \code{num_features} using all features present in object.
\item Run \code{AddModuleScore} for all of those feature lists.
\item Extract the median and calculate median + 3xMAD for each of the 1000 scores.
\item Plot median + 3xMAD distribution.
\item Extract 97.5\% quantile value and save in misc slot.
\item Calculate and store data.frame with the percent of cells with \code{score_name} above
calculated threshold.
\item Plot \code{score_name} by identity with line indicating threshold and labels indicating the
percent enriched in each identity.  Will store plots in misc if \code{return_plots = TRUE}.
}
}
\examples{
\dontrun{
obj <- Run_Module_Sig(seurat_object = obj, score_name = "module_score", num_features = 25,
downsample_cell_num = 2000, return_plots = TRUE)
}

}
\references{
Marsh et al., (2022). Dissection of artifactual and confounding glial signatures by
single-cell sequencing of mouse and human brain. Nat Neurosci, 25(3):306–316. PMID: 35260865.
\url{doi.org/10.1038/s41593-022-01022-8}.
}
\concept{stats}
