% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_QC_Seq_10X.R
\name{Seq_QC_Plot_Genome}
\alias{Seq_QC_Plot_Genome}
\title{QC Plots Sequencing metrics (Alignment)}
\usage{
Seq_QC_Plot_Genome(
  metrics_dataframe,
  group.by = "sample_id",
  plot_by = deprecated(),
  colors_use = NULL,
  dot_size = 1,
  x_lab_rotate = FALSE,
  significance = FALSE,
  ...
)
}
\arguments{
\item{metrics_dataframe}{data.frame  contain Cell Ranger QC Metrics (see \code{\link{Read_Metrics_10X}}).}

\item{group.by}{Grouping factor for the plot.  Default is to plot as single group with single point per sample.}

\item{plot_by}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} soft-deprecated. See \code{group.by}.}

\item{colors_use}{colors to use for plot if plotting by group.  Defaults to RColorBrewer Dark2 palette if
less than 8 groups and \code{DiscretePalette_scCustomize(palette = "polychrome")} if more than 8.}

\item{dot_size}{size of the dots plotted if \code{plot_by} is not \code{sample_id}  Default is 1.}

\item{x_lab_rotate}{logical.  Whether to rotate the axes labels on the x-axis.  Default is FALSE.}

\item{significance}{logical.  Whether to calculate and plot p-value comparisons when plotting by
grouping factor.  Default is FALSE.}

\item{...}{Other variables to pass to \code{ggpubr::stat_compare_means} when doing significance testing.}
}
\value{
A ggplot object
}
\description{
Plot the fraction of reads confidently mapped to genome
}
\examples{
\dontrun{
Seq_QC_Plot_Genome(metrics_dataframe = metrics)
}

}
\concept{seq_qc_plotting_alignment}
