% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Setup_scRNAseq_Project}
\alias{Setup_scRNAseq_Project}
\title{Setup project directory structure}
\usage{
Setup_scRNAseq_Project(
  custom_dir_file = NULL,
  cluster_annotation_path = NULL,
  cluster_annotation_file_name = "cluster_annotation.csv"
)
}
\arguments{
\item{custom_dir_file}{file to file containing desired directory structure.  Default is NULL and
will provide generic built-in directory structure.}

\item{cluster_annotation_path}{path to place cluster annotation file using \code{\link{Create_Cluster_Annotation_File}}.}

\item{cluster_annotation_file_name}{name to use for annotation file if created (optional).}
}
\value{
no return value.  Creates system folders.
}
\description{
Create reproducible project directory organization when initiating a new analysis.
}
\examples{
\dontrun{
# If using built-in directory structure.
Setup_scRNAseq_Project()
}

}
\concept{organization_util}
