% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object_Conversion.R
\name{Split_Layers}
\alias{Split_Layers}
\title{Split Seurat object into layers}
\usage{
Split_Layers(seurat_object, assay = "RNA", split.by)
}
\arguments{
\item{seurat_object}{Seurat object name.}

\item{assay}{name(s) of assays to convert.  Defaults to current active assay.}

\item{split.by}{Variable in meta.data to use for splitting layers.}
}
\description{
Split Assay5 of Seurat object into layers by variable in meta.data
}
\examples{
\dontrun{
# Split object by "treatment"
obj <- Split_Layers(object = obj, assay = "RNA", split.by = "treatment")
}

}
\concept{object_conversion}
