% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_Plotting.R
\name{Stacked_VlnPlot}
\alias{Stacked_VlnPlot}
\title{Stacked Violin Plot}
\usage{
Stacked_VlnPlot(
  seurat_object,
  features,
  group.by = NULL,
  split.by = NULL,
  idents = NULL,
  x_lab_rotate = FALSE,
  colors_use = NULL,
  color_seed = 123,
  ggplot_default_colors = FALSE,
  ...
)
}
\arguments{
\item{seurat_object}{Seurat object name.}

\item{features}{Features to plot.}

\item{group.by}{Group (color) cells in different ways (for example, orig.ident).}

\item{split.by}{A variable to split the violin plots by,}

\item{idents}{Which classes to include in the plot (default is all).}

\item{x_lab_rotate}{Rotate x-axis labels 45 degrees (Default is FALSE).}

\item{colors_use}{specify color palette to used in \code{\link[Seurat]{VlnPlot}}.  By default if
number of levels plotted is less than or equal to 36 it will use "polychrome" and if greater than 36
will use "varibow" with shuffle = TRUE both from \code{DiscretePalette_scCustomize}.}

\item{color_seed}{random seed for the "varibow" palette shuffle if \code{colors_use = NULL} and number of
groups plotted is greater than 36.  Default = 123.}

\item{ggplot_default_colors}{logical.  If \code{colors_use = NULL}, Whether or not to return plot using
default ggplot2 "hue" palette instead of default "polychrome" or "varibow" palettes.}

\item{...}{Extra parameters passed to \code{\link[Seurat]{VlnPlot}}.}
}
\value{
A ggplot object
}
\description{
Code for creating stacked violin plot gene expression.
}
\examples{
\dontrun{
Stacked_VlnPlot(seurat_object = object, features = gene_list, x_lab_rotate = TRUE)
}

}
\references{
https://divingintogeneticsandgenomics.rbind.io/post/stacked-violin-plot-for-visualizing-single-cell-data-in-seurat/
}
\seealso{
https://twitter.com/tangming2005
}
\author{
Ming Tang (Original Code), Sam Marsh (Wrap single function, added/modified functionality)
}
\concept{seurat_plotting}
