% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Updated_HGNC_Symbols}
\alias{Updated_HGNC_Symbols}
\title{Update HGNC Gene Symbols}
\usage{
Updated_HGNC_Symbols(input_data, update_symbol_data = NULL, verbose = TRUE)
}
\arguments{
\item{input_data}{Data source containing gene names.  Accepted formats are:
\itemize{
\item \code{charcter vector}
\item \code{Seurat Objects}
\item \code{data.frame}: genes as rownames
\item \code{dgCMatrix/dgTMatrix}: genes as rownames
\item \code{tibble}: genes in first column
}}

\item{update_symbol_data}{logical, whether to update cached HGNC data, default is NULL.
If \code{NULL} BiocFileCache will check and prompt for update if cache is stale.
If \code{FALSE} the BiocFileCache stale check will be skipped and current cache will be used.
If \code{TRUE} the BiocFileCache stale check will be skipped and HGNC data will be downloaded.}

\item{verbose}{logical, whether to print results detailing numbers of symbols, found, updated,
and not found; default is TRUE.}
}
\value{
data.frame containing columns: input_features, Approved_Symbol (already approved; output unchanged), Not_Found_Symbol (symbol not in HGNC; output unchanged), Updated_Symbol (new symbol from HGNC; output updated).
}
\description{
Update human gene symbols using data from HGNC.
}
\examples{
\dontrun{
# Function supports various inputs including: Seurat Objects, character vectors, data.frame, and sparse matrices (with genes as rows)
new_names <- Updated_HGNC_Symbols(input_data = Seurat_Object)
}

}
\concept{helper_util}
