% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Color_Palettes.R
\docType{data}
\name{viridis_dark_high}
\alias{viridis_dark_high}
\title{Viridis Shortcuts}
\format{
An object of class \code{character} of length 10.
}
\usage{
viridis_dark_high
}
\description{
Quick shortcuts to access viridis palettes
}
\examples{
\dontrun{
FeaturePlot_scCustom(object = seurat_object, features = "Cx3cr1", colors_use = viridis_plasma_dark_high, na_color = "lightgray")
}

}
\keyword{datasets}
