% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Color_Palettes.R
\name{viridis_plasma_dark_high}
\alias{viridis_plasma_dark_high}
\alias{viridis_plasma_light_high}
\alias{viridis_inferno_dark_high}
\alias{viridis_inferno_light_high}
\alias{viridis_magma_dark_high}
\alias{viridis_magma_light_high}
\alias{viridis_dark_high}
\alias{viridis_light_high}
\title{Viridis Shortcuts}
\usage{
viridis_plasma_dark_high

viridis_plasma_light_high

viridis_inferno_dark_high

viridis_inferno_light_high

viridis_magma_dark_high

viridis_magma_light_high

viridis_dark_high

viridis_light_high
}
\value{
A color palette for plotting
}
\description{
Quick shortcuts to access viridis palettes
}
\examples{
\dontrun{
FeaturePlot_scCustom(object = seurat_object, features = "Cx3cr1",
colors_use = viridis_plasma_dark_high, na_color = "lightgray")
}

}
\concept{palettes}
