% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIGER_Utilities.R
\name{WhichCells.liger}
\alias{WhichCells.liger}
\title{Extract Cells for particular identity}
\usage{
\method{WhichCells}{liger}(object, idents = NULL, by_dataset = FALSE, invert = FALSE, ...)
}
\arguments{
\item{idents}{identities to extract cell barcodes}

\item{by_dataset}{logical, whether to return vector with cell barcodes for all \code{idents} in or
to return list (1 entry per dataset with vector of cells) (default is FALSE; return vector).}

\item{invert}{logical, invert the selection of cells (default is FALSE).}

\item{liger_object}{LIGER object name.}
}
\value{
vector or list depending on \code{by_dataset} parameter
}
\description{
Extract all cell barcodes for a specific identity
}
\examples{
\dontrun{
ident1_cells <- WhichCells(object = liger_object, idents = 1)
}

}
\concept{object_conversion}
