% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object_Utilities.R
\name{Add_Alt_Feature_ID}
\alias{Add_Alt_Feature_ID}
\title{Add Alternative Feature IDs}
\usage{
Add_Alt_Feature_ID(
  seurat_object,
  features_tsv_file = NULL,
  hdf5_file = NULL,
  assay = NULL
)
}
\arguments{
\item{seurat_object}{object name.}

\item{features_tsv_file}{output file from Cell Ranger used for creation of Seurat object.
(Either provide this of \code{hdf5_file})}

\item{hdf5_file}{output file from Cell Ranger used for creation of Seurat object.
(Either provide this of \code{features_tsv_file})}

\item{assay}{name of assay(s) to add the alternative features to.  Can specify "all"
to add to all assays.}
}
\value{
Seurat Object with new entries in the \code{obj@assays$ASSAY@meta.data} slot.
}
\description{
Add alternative feature ids to the assay level meta.data slot in Assay5 compatible object (Seurat V5.0.0 or greater)
}
\examples{
\dontrun{
test_obj <- Add_Alt_Feature_ID(seurat_object = test_obj,
features_tsv = "/FILE_PATH/features.tsv.gz", assay = "RNA")
}

}
\concept{get_set_util}
