% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/Object_Conversion.R
\name{as.anndata}
\alias{as.anndata}
\alias{as.anndata.Seurat}
\title{Convert objects to anndata objects}
\usage{
as.anndata(x, ...)

\method{as.anndata}{Seurat}(
  x,
  file_path,
  file_name,
  assay = "RNA",
  main_layer = "data",
  other_layers = "counts",
  transer_dimreduc = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Seurat or LIGER object}

\item{...}{Arguments passed to other methods}

\item{file_path}{directory file path and/or file name prefix.  Defaults to current wd.}

\item{file_name}{file name.}

\item{assay}{Assay containing data to use, (default is "RNA").}

\item{main_layer}{the layer of data to become default layer in anndata object (default is "data").}

\item{other_layers}{other data layers to transfer to anndata object (default is "counts").}

\item{transer_dimreduc}{logical, whether to transfer dimensionality reduction coordinates from
Seurat to anndata object (default is TRUE).}

\item{verbose}{logical, whether to print status messages during object conversion (default is TRUE).}
}
\value{
saved anndata object to at path provided.
}
\description{
This function is part of generic \code{as.anndata} for conversion of Seurat Objects to anndata objects.
}
\examples{
\dontrun{
as.anndata(x = seurat_object, file_path = "/folder_name", file_name = "anndata_converted.h5ad")
}

}
\references{
modified and enhanced version of \code{sceasy::seurat2anndata} (sceasy package: \url {https://github.com/cellgeni/sceasy}; License: GPL-3.  Function has additional checks and supports Seurat V3 and V5 object structure.
}
\concept{object_conversion}
