% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{as.Seurat}
\alias{as.Seurat}
\title{Convert objects to Seurat objects}
\usage{
as.Seurat(x, ...)
}
\arguments{
\item{x}{An object(s) to convert to class \code{Seurat}}

\item{...}{Arguments passed to other methods}
}
\description{
Convert objects to Seurat objects
}
