% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object_Conversion.R
\name{as.Seurat.liger}
\alias{as.Seurat.liger}
\title{Convert to Seurat Object}
\usage{
\method{as.Seurat}{liger}(
  x,
  nms = names(liger_object@H),
  renormalize = TRUE,
  use.liger.genes = TRUE,
  by.dataset = FALSE,
  keep_meta = TRUE,
  reduction_label = "UMAP",
  seurat_assay = "RNA",
  assay_type = NULL,
  add_barcode_names = FALSE,
  barcode_prefix = TRUE,
  barcode_cell_id_delimiter = "_",
  ...
)
}
\value{
Seurat object.
}
\description{
Convert to Seurat Object
}
\examples{
\dontrun{
seurat_object <- as.Seurat(x = liger_object)
}

}
\concept{object_conversion}
