% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object_Conversion.R
\name{as.Seurat.liger}
\alias{as.Seurat.liger}
\title{Convert to Seurat Object}
\usage{
\method{as.Seurat}{liger}(x, meta_data = TRUE, ...)
}
\value{
Seurat object.
}
\description{
Convert to Seurat Object
}
\examples{
\dontrun{
liger_object <- as.Seurat(x = seurat_object_list)
}

}
\concept{object_conversion}
