% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Change_Delim_Prefix}
\alias{Change_Delim_Prefix}
\title{Change barcode prefix delimiter}
\usage{
Change_Delim_Prefix(data, current_delim, new_delim)
}
\arguments{
\item{data}{Either matrix/data.frame or list of matrices/data.frames with the cell barcodes in the column names.}

\item{new_delim}{a single value of new delimiter desired.}

\item{currnet_delim}{a single value of current delimiter.}
}
\description{
Change barcode prefix delimiter from list of data.frames/matrices or single data.frame/matrix
}
\examples{
\dontrun{
dge_matrix <- Change_Delim_Prefix(data = dge_matrix, currnet_delim = ".", new_delim = "-")
}

}
\concept{helper_util}
