% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_Plotting.R
\name{Cluster_Highlight_Plot}
\alias{Cluster_Highlight_Plot}
\title{Cluster Highlight Plot}
\usage{
Cluster_Highlight_Plot(
  seurat_object,
  cluster_name,
  highlight_color = "navy",
  background_color = "lightgray",
  pt.size = NULL,
  raster = NULL,
  ...
)
}
\arguments{
\item{seurat_object}{Seurat object name.}

\item{cluster_name}{Name(s) (or number(s)) identity of cluster to be highlighted.}

\item{highlight_color}{Color(s) to highlight cells (default "navy").}

\item{background_color}{non-highlighted cell colors.}

\item{pt.size}{point size for both highlighted cluster and background.}

\item{raster}{Convert points to raster format.  Default is NULL which will rasterize by default if
greater than 200,000 cells.}

\item{...}{Extra parameters passed to \code{\link[Seurat]{DimPlot}}.}
}
\value{
A ggplot object
}
\description{
Create Plot with cluster of interest highlighted
}
\examples{
\dontrun{
Cluster_Highlight_Plot(seurat_object = object, cluster_name = "Microglia", highlight_color = "gold",
background_color = "lightgray",  pt.size = 2)
}

}
\concept{seurat_plotting}
