% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{Cluster_Stats_All_Samples}
\alias{Cluster_Stats_All_Samples}
\title{Calculate Cluster Stats}
\usage{
Cluster_Stats_All_Samples(seurat_object, group_by_var = "orig.ident")
}
\arguments{
\item{seurat_object}{Seurat object name.}

\item{group_by_var}{meta data column to classify samples (default = "orig.ident").}
}
\value{
A data.frame with rows in order of frequency
}
\description{
Calculates both overall and per sample cell number and percentages per cluster based on orig.ident.
}
\examples{
\dontrun{
stats <- Cluster_Stats_All_Samples(seurat_object = object, group_by_var = "orig.ident")
}

}
\concept{stats}
