% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIGER_Utilities.R
\name{Dataset_Size_LIGER}
\alias{Dataset_Size_LIGER}
\title{Check size of LIGER datasets}
\usage{
Dataset_Size_LIGER(
  liger_object,
  meta_data_column = NULL,
  filter_by = NULL,
  print_filter = FALSE
)
}
\arguments{
\item{liger_object}{LIGER object name.}

\item{meta_data_column}{other meta data to include in returned data.frame.}

\item{filter_by}{meta data column to filter data by.  Will filter data to return only values for  the
largest dataset for each unique value in provided meta data column.}

\item{print_filter}{logical, whether to print filtered results to console, default is FALSE.}
}
\value{
data.frame with dataset names, number of cells per dataset and if provided other meta data
}
\description{
Returns size (number of cells) in each dataset within liger object along with other desired meta data.
}
\examples{
\dontrun{
# Return values for all datasets
}

}
\concept{liger_object_util}
