% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{Add_Mito_Ribo_Seurat}
\title{Deprecated functions}
\usage{
Add_Mito_Ribo_Seurat(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Use \code{\link[=Add_Mito_Ribo]{Add_Mito_Ribo()}} instead of \code{Add_Mito_Ribo_Seurat()}.
}
\keyword{internal}
