% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_Seurat.R
\name{ElbowPlot_scCustom}
\alias{ElbowPlot_scCustom}
\title{ElbowPlot with modifications}
\usage{
ElbowPlot_scCustom(
  seurat_object,
  ndims = NULL,
  reduction = "pca",
  calc_cutoffs = TRUE,
  plot_cutoffs = TRUE,
  line_colors = c("blue", "red"),
  linewidth = 0.5
)
}
\arguments{
\item{seurat_object}{name of Seurat object}

\item{ndims}{The number of dims to plot.  Default is NULL and will plot all dims}

\item{reduction}{The reduction to use, default is "pca"}

\item{calc_cutoffs}{logical, whether or not to calculate the cutoffs}

\item{linewidth}{}
}
\value{
ggplot2 object
}
\description{
Creates ElbowPlot with ability to calculate and plot cutoffs for amount of variance conferred by the PCs.
Cutoff 1 is PC where principal components only contribute 5\% of standard deviation and the principal components
cumulatively contribute 90\% of the standard deviation.
Cutoff 2 is point where the percent change in variation between the consecutive PCs is less than 0.1\%.
}
\references{
Modified from following: \url{https://hbctraining.github.io/scRNA-seq/lessons/elbow_plot_metric.html}
}
