% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_Plotting.R
\name{FeaturePlot_scCustom}
\alias{FeaturePlot_scCustom}
\title{Customize FeaturePlot}
\usage{
FeaturePlot_scCustom(
  seurat_object,
  features,
  colors_use = viridis_plasma_dark_high,
  na_color = "lightgray",
  order = TRUE,
  pt.size = NULL,
  reduction = NULL,
  na_cutoff = 1e-09,
  raster = NULL,
  split.by = NULL,
  num_columns = NULL,
  slot = "data",
  alpha_exp = NULL,
  alpha_na_exp = NULL,
  ...
)
}
\arguments{
\item{seurat_object}{Seurat object name.}

\item{features}{Feature(s) to plot.}

\item{colors_use}{list of colors or color palette to use.}

\item{na_color}{color to use for points below lower limit.}

\item{order}{whether to move positive cells to the top (default = TRUE).}

\item{pt.size}{Adjust point size for plotting.}

\item{reduction}{Dimensionality Reduction to use (if NULL then defaults to Object default).}

\item{na_cutoff}{Value to use as minimum expression cutoff.  This will be lowest value plotted use
palette provided to \code{colors_use}.  Leave as default value to plot only positive non-zero values using
color scale and zero/negative values as NA.  To plot all values using color palette set to \code{NA}.}

\item{raster}{Convert points to raster format.  Default is NULL which will rasterize by default if
greater than 200,000 cells.}

\item{split.by}{Variable in \verb{@meta.data} to split the plot by.}

\item{num_columns}{Number of columns in plot layout.}

\item{slot}{Which slot to pull expression data from?  Default is "data".}

\item{alpha_exp}{new alpha level to apply to expressing cell color palette (\code{colors_use}).  Must be
value between 0-1.}

\item{alpha_na_exp}{new alpha level to apply to non-expressing cell color palette (\code{na_color}).  Must be
value between 0-1.}

\item{...}{Extra parameters passed to \code{\link[Seurat]{FeaturePlot}}.}
}
\value{
A ggplot object
}
\description{
Create Custom FeaturePlots and preserve scale (no binning)
}
\examples{
\dontrun{
FeaturePlot_scCustom(seurat_object = object, features = "Cx3cr1", colors_use = viridis_plasma_dark_high,
na_color = "lightgray")
}

}
\concept{seurat_plotting}
