% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Fetch_Meta}
\alias{Fetch_Meta}
\alias{Fetching_Meta.Seurat}
\alias{Fetching_Meta.liger}
\title{Get meta data from object}
\usage{
Fetch_Meta(object)

Fetching_Meta.Seurat(object)

Fetching_Meta.liger(object)
}
\arguments{
\item{object}{list of matrices to merge.}
}
\value{
A data.frame
}
\description{
Quick function to properly pull meta.data from objects.
}
\examples{
\dontrun{
meta_data <- Fetch_Meta(object = pbmc)
}

}
\concept{helper_util}
