% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/LIGER_Utilities.R,
%   R/Object_Utilities.R
\name{Fetch_Meta}
\alias{Fetch_Meta}
\alias{Fetch_Meta.liger}
\alias{Fetch_Meta.Seurat}
\title{Get meta data from object}
\usage{
Fetch_Meta(object, ...)

\method{Fetch_Meta}{liger}(object, ...)

\method{Fetch_Meta}{Seurat}(object, columns = NULL, ...)
}
\arguments{
\item{object}{Object of class Seurat or liger.}

\item{...}{Arguments passed to other methods}

\item{columns}{optional, name(s) of columns to return. Default is NULL; returns all columns}
}
\value{
A data.frame containing cell-level meta data
}
\description{
Quick function to properly pull meta.data from objects.
}
\examples{
library(Seurat)
meta_data <- Fetch_Meta(object = pbmc_small)
head(meta_data, 5)

}
\concept{get_set_util}
\concept{liger_object_util}
