% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Fetch_Meta}
\alias{Fetch_Meta}
\title{Get meta data from object}
\usage{
Fetch_Meta(object)
}
\arguments{
\item{object}{Seurat object}
}
\value{
A data.frame
}
\description{
Quick function to properly pull meta.data from objects.
}
\examples{
\dontrun{
meta_data <- Fetch_Meta(object = pbmc)
}

}
\concept{helper_util}
