% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIGER_Utilities.R
\name{Find_Factor_Cor}
\alias{Find_Factor_Cor}
\title{Find Factor Correlations}
\usage{
Find_Factor_Cor(object, reduction)
}
\arguments{
\item{object}{LIGER/Seurat object name.}

\item{reduction}{reduction name to pull loadings for.  Only valid if supplying a Seurat object.}
}
\value{
correlation matrix
}
\description{
Calculate correlations between gene loadings for all factors in liger or Seurat object.
}
\examples{
\dontrun{
factor_correlations <- Find_Factor_Cor(object = object)
}

}
\concept{liger_object_util}
