% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_Utilities.R
\name{Move_Legend}
\alias{Move_Legend}
\title{Move Legend Position}
\usage{
Move_Legend(position = "right", ...)
}
\arguments{
\item{position}{valid position to move legend.  Default is "right".}

\item{...}{extra arguments passed to \code{ggplot2::theme()}.}
}
\value{
Returns a list-like object of class \emph{theme}.
}
\description{
Shortcut for thematic modification to move legend position.
}
\examples{
\dontrun{
# Generate a plot and move the legend.
library(ggplot2)
p <- VlnPlot(object = obj, features = "Cx3cr1")
p + Move_Legend("left")
}

}
\concept{themes}
