% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Color_Palettes.R
\name{PalettePlot}
\alias{PalettePlot}
\title{Plot color palette in viewer}
\usage{
PalettePlot(palette = NULL)
}
\arguments{
\item{palette}{a vector of colors (either named colors of hex codes).}
}
\value{
Plot of all colors in supplied palette/vector
}
\description{
Plots given color vector/palette in viewer to evaluate palette before plotting on data.
}
\examples{
pal <- DiscretePalette_scCustomize(num_colors = 36, palette = "varibow")
PalettePlot(palette = pal)

}
\references{
Adapted from colorway package \code{build_palette} internals (License: GPL-3).
\url{https://github.com/hypercompetent/colorway}.
}
\concept{palettes}
