% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as.Seurat}
\alias{WhichCells}
\alias{Cells}
\alias{Features}
\alias{Embeddings}
\title{Objects exported from other packages}
\note{
See \code{\link{as.Seurat.liger}} for scCustomize extension of this generic to converting Liger objects.

See \code{\link{WhichCells.liger}} for scCustomize extension of this generic to extract cell barcodes.

See \code{\link{Cells.liger}} for scCustomize extension of this generic to extract cell barcodes.

See \code{\link{Features.liger}} for scCustomize extension of this generic to extract dataset features.

See \code{\link{Embeddings.liger}} for scCustomize extension of this generic to extract embeddings.
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{SeuratObject}{\code{\link[SeuratObject]{as.Seurat}}, \code{\link[SeuratObject]{Cells}}, \code{\link[SeuratObject]{Embeddings}}, \code{\link[SeuratObject:Cells]{Features}}, \code{\link[SeuratObject]{WhichCells}}}
}}

