% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as.Seurat}
\title{Objects exported from other packages}
\note{
See \code{as.Seurat.liger()} for scCustomize extension of this generic to converting Liger objects.
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{SeuratObject}{\code{\link[SeuratObject]{as.Seurat}}}
}}

