% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Rename_Clusters}
\alias{Rename_Clusters}
\title{Rename Cluster Seurat}
\usage{
Rename_Clusters(
  seurat_object,
  new_idents,
  old_ident_name = NULL,
  new_ident_name = NULL,
  meta_col_name = deprecated(),
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{seurat_object}{object name.}

\item{new_idents}{vector of new cluster names.  Must be equal to the length of current active.ident
in Seurat Object.  Will accept named vector (with old idents as names) or will name the new_idents vector internally.}

\item{old_ident_name}{optional, name to use for storing current object idents in \code{Object@meta.data}.}

\item{new_ident_name}{optional, name to use for storing new object idents in \verb{@meta.data}.}

\item{meta_col_name}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#soft-deprecated}{\figure{lifecycle-soft-deprecated.svg}{options: alt='[Soft-deprecated]'}}}{\strong{[Soft-deprecated]}}. See \code{old_ident_name}.}

\item{overwrite}{logical, whether to overwrite columns in \verb{@meta.data} if they have same
names as \code{old_ident_name} and/or \code{new_ident_name}.}

\item{...}{Extra parameters passed to \code{\link[SeuratObject]{RenameIdents}}.}
}
\value{
Seurat Object with new identities placed in active.ident slot.
}
\description{
Wrapper function to rename active identities in Seurat Object with new idents.
}
\examples{
\dontrun{
obj <- Rename_Clusters(seurat_object = obj_name, new_idents = new_idents_vec,
old_ident_name = "Seurat_Idents_Round01", new_ident_name = "Round01_Res0.6_Idents")
}

}
\concept{marker_annotation_util}
