% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIGER_Utilities.R
\name{Top_Genes_Factor}
\alias{Top_Genes_Factor}
\title{Extract top loading genes for LIGER factor}
\usage{
Top_Genes_Factor(liger_object, liger_factor, num_genes = 10)
}
\arguments{
\item{liger_object}{LIGER object name.}

\item{liger_factor}{LIGER factor number to pull genes from.}

\item{num_genes}{number of top loading genes to return as vector.}
}
\value{
A LIGER Object
}
\description{
Extract vector to the top loading genes for specified LIGER iNMF factor
}
\examples{
\dontrun{
top_genes_factor10 <- Top_Genes_Factor(liger_object = object, num_genes = 10)
}

}
\concept{liger_object_util}
