% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_Utilities.R
\name{UnRotate_X}
\alias{UnRotate_X}
\title{Unrotate x axis on VlnPLot}
\usage{
UnRotate_X(...)
}
\arguments{
\item{...}{extra arguments passed to \code{ggplot2::theme()}.}
}
\value{
Returns a list-like object of class \emph{theme}.
}
\description{
Shortcut for thematic modification to unrotate the x axis (e.g., for Seurat VlnPlot is rotated by default).
}
\examples{
library(Seurat)
p <- VlnPlot(object = pbmc_small, features = "CD3E")
p + UnRotate_X

}
\concept{themes}
