% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_Utilities.R
\name{UnRotate_X}
\alias{UnRotate_X}
\title{Unrotate x axis on VlnPLot}
\usage{
UnRotate_X(...)
}
\description{
Shortcut for thematic modification to unrotate the x axis (e.g., for Seurat VlnPlot is rotated by default).
}
\examples{
\dontrun{
# Generate a plot and unrotate the x-axis label
library(ggplot2)
p <- VlnPlot(object = obj, features = "Cx3cr1")
p + UnRotate_X
}

}
\concept{themes}
